actuate: dshot: Update node to account for a minimum update-rate
Consider the use-case where the actuator updates are not coming at a
fast-enough pace to drive the DShot appropriately. In this case, have
the default update-rate be the worst-case update rate.

Alternative solution, inspired by: 6e49452.

Disclaimer: this has not be tested, so don't assume this is a final
solution, but rather a conceptual approach.

Signed-off-by: Luis Ubieda <luisf@croxel.com>