tree-wide: Offload ZROS sense drivers to external repository (#236)
* manifest: Add zros-drivers to the list of submodules

As a first step to offload common drivers to this module. No functional
changes at this point.

Signed-off-by: Luis Ubieda <luisf@croxel.com>

* zros: offload perf utilities to zros-drivers repo

Updated manifest to go in sync with changes.

Signed-off-by: Luis Ubieda <luisf@croxel.com>

* all: Remove vestigial perf dependencies

There are a few modules including it but not making use of it.
Scrub them.

Signed-off-by: Luis Ubieda <luisf@croxel.com>

* sense: imu: Move driver to zros_drivers repo

In order to be able to use it in other apps.
In this case the user is required to know the CEREBRI_SENSE_IMU is
now ZROS_SENSE_IMU.

Signed-off-by: Luis Ubieda <luisf@croxel.com>

* sense: baro: Move to zros_drivers

For common driver usage.

Signed-off-by: Luis Ubieda <luisf@croxel.com>

* sense: mag: move to zros_drivers

To use in other apps.

Signed-off-by: Luis Ubieda <luisf@croxel.com>

---------

Signed-off-by: Luis Ubieda <luisf@croxel.com>