Enable CODE_DATA_RELOCATION when using ITCM
When relocating code to ITCM, there's a nasty bug in zephyr that
discarded functions still end up in the ELF at address 0x0
Functionality wise this is not a problem but when using a debugger
it can't make sense of the first ~0x100 bytes of the ITCM

As it turns out the CODE_DATA_RELOCATION config somehow solves this
However it requires that you use zephyr_code_relocate atleast once.
Looking at basic heuristics it seems we want nanopb in ITCM anyhow
So instead of the manual ld we use the zephyr_code_relocate to relocate
the text secionts of the nanopb library.

Signed-off-by: Peter van der Perk <peter.vanderperk@nxp.com>